requirements.txt
discord.py==2.6.3
requests>=2.32.5
beautifulsoup4>=4.13.5
pandas>=2.1.0
redis>=5.0.0
rapidfuzz>=2.16.0
